CloudMaxFileSize type changed to int64 (issue #213)

+++ Settings.pas
@@ -107,7 +107,7 @@ type
 		DescriptionTrackCloudFS: boolean;
 		DescriptionFileName: WideString;
 		CopyBetweenAccountsMode: integer;
-		CloudMaxFileSize: integer;
+		CloudMaxFileSize: int64;
 		ChunkOverwriteMode: integer;
 		DeleteFailOnUploadMode: integer;
 		OperationErrorMode: integer;
